requirements.txt
grpcio==1.47.0
grpcio-tools==1.47.0
Flask==2.2.2